+++ Dockerfile
@@ -10,4 +10,5 @@ RUN npm install
 COPY . .
 
 EXPOSE 8080
+
 CMD [ "npm", "start" ]
