+++ src/handlers/authentication.js
@@ -1,0 +1,117 @@
+const DAL = require('../dal/dal');
+const JWT = require('../shared/jwt');
+
+/**
+ * @swagger
+ * /auth/:
+ *  post:
+ *    description: Use to sign up a user
+ *    parameters:
+ *      - in: body
+ *        required: true
+ *        schema:
+ *          type: object
+ *          properties:
+ *            firstName:
+ *              type: string
+ *            lastName:
+ *              type: string
+ *            email:
+ *              type: string
+ *            password:
+ *              type: string
+ *    responses:
+ *      '201':
+ *        description: Success
+ *        schema:
+ *          type: object
+ *          properties:
+ *            user:
+ *              type: object
+ *            token:
+ *              type: string
+ *      '400':
+ *        description: 'Bad request'
+ *      '500':
+ *        description: 'Internal server error'
+ */
+let signUp = async (req, res) => {
+    let firstName = req.body.firstName,
+        lastName = req.body.lastName,
+        email = req.body.email,
+        password = req.body.password;
+
+    if (firstName === undefined || lastName === undefined || email === undefined || password === undefined) {
+        return res.status(400).send();
+    }
+
+    let signUpRes = await DAL.signUp(firstName, lastName, email, password);
+    if (signUpRes.err) {
+        return res.status(signUpRes.err).send();
+    }
+
+    let token = JWT.issueUserToken(signUpRes.user.uid);
+
+    return res.status(201).send({user: signUpRes.user, token: token});
+};
+
+/**
+ * @swagger
+ * /auth/:
+ *  put:
+ *    description: Use to sign in a user
+ *    parameters:
+ *      - in: body
+ *        required: true
+ *        schema:
+ *          type: object
+ *          properties:
+ *            email:
+ *              type: string
+ *            password:
+ *              type: string
+ *    responses:
+ *      '200':
+ *        description: Success
+ *        schema:
+ *          type: object
+ *          properties:
+ *            user:
+ *              type: object
+ *            token:
+ *              type: string
+ *      '400':
+ *        description: 'Bad request'
+ *      '404':
+ *        description: 'Not found'
+ *      '500':
+ *        description: 'Internal server error'
+ */
+let signIn = async (req, res) => {
+    let email = req.body.email,
+        password = req.body.password;
+
+    if (email === undefined || password === undefined) {
+        return res.status(400).send('must provide email and password');
+    }
+
+    let signInRes = await DAL.signInEmailPassword(email, password);
+    if (signInRes.err) {
+        return res.status(signInRes.err).send();
+    }
+
+    let token;
+    if (signInRes.user.admin) {
+        token = JWT.issueAdminToken(signInRes.user.uid);
+    }
+    else {
+        token = JWT.issueUserToken(signInRes.user.uid);
+    }
+
+    return res.status(200).send({user: signInRes.user, token: token});
+};
+
+module.exports = {
+    signUp,
+    signIn
+};
