+++ src/handlers/registration.js
@@ -1,0 +1,152 @@
+/**
+ * @swagger
+ * /hacks/:hid/reg/q/:
+ *  post:
+ *    description: Use to create a registration question
+ *    parameters:
+ *      - name: ha-admin-token
+ *        in: header
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *      - name: hid
+ *        in: path
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *    responses:
+ *      '200':
+ *        description: Success
+ *      '400':
+ *        description: 'Bad request'
+ *      '401':
+ *        description: 'JWT not found in header'
+ *      '403':
+ *        description: 'JWT does not have admin privileges'
+ */
+let createRegQuestion = async (req, res) => {
+
+};
+
+/**
+ * @swagger
+ * /hacks/:hid/reg/q/:qid/:
+ *  put:
+ *    description: Use to update a registration question
+ *    parameters:
+ *      - name: ha-admin-token
+ *        in: header
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *      - name: hid
+ *        in: path
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *      - name: qid
+ *        in: path
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *    responses:
+ *      '200':
+ *        description: Success
+ *      '400':
+ *        description: 'Bad request'
+ *      '401':
+ *        description: 'JWT not found in header'
+ *      '403':
+ *        description: 'JWT does not have admin privileges'
+ *      '404':
+ *        description: 'Not found'
+ */
+let updateRegQuestion = async (req, res) => {
+
+};
+
+/**
+ * @swagger
+ * /hacks/:hid/reg/q/:qid/:
+ *  delete:
+ *    description: Use to delete a registration question
+ *    parameters:
+ *      - name: ha-admin-token
+ *        in: header
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *      - name: hid
+ *        in: path
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *      - name: qid
+ *        in: path
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *    responses:
+ *      '200':
+ *        description: Success
+ *      '400':
+ *        description: 'Bad request'
+ *      '401':
+ *        description: 'JWT not found in header'
+ *      '403':
+ *        description: 'JWT does not have admin privileges'
+ *      '404':
+ *        description: 'Not found'
+ */
+let deleteRegQuestion = async (req, res) => {
+
+};
+
+/**
+ * @swagger
+ * /hacks/:hid/reg/users/csv/:
+ *  get:
+ *    description: Get a user registration form
+ *    parameters:
+ *      - name: ha-admin-token
+ *        in: header
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *      - name: hid
+ *        in: path
+ *        required: true
+ *        schema:
+ *          type: string
+ *          format: uuid
+ *    responses:
+ *      '200':
+ *        description: Success
+ *      '400':
+ *        description: 'Bad request'
+ *      '401':
+ *        description: 'JWT not found in header'
+ *      '403':
+ *        description: 'JWT does not have admin privileges'
+ *      '404':
+ *        description: 'Not found'
+ */
+let getUserRegForm = async (req, res) => {
+
+};
+
+module.exports = {
+    createRegQuestion: createRegQuestion,
+    updateRegQuestion: updateRegQuestion,
+    deleteRegQuestion: deleteRegQuestion,
+    getUserRegForm
+};
